requirements.txt
streamlit
pandas
requests
python-binance
plotly